api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
pyjwt==2.8.0
stripe==9.9.0
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.2.1
python-multipart==0.0.9
pydantic==2.7.1
